Properly Send Optional carrier_accounts Parameter When Creating Shipment  (#4)
### What It Fixes ###
The _CreateShipment_ method supports an optional carrier account ID parameter. If specified along with providing a service on the shipment object, shipments can be created and bought with one request.

Previously, we were not sending the optional carrier account parameter in the right way, which caused the API to return an error. This change fixes that bug.

### How I Tested It ###
I re-ran existing automated tests and made sure they pass. I did not want to add an automated test for the "create and buy" functionality, as this requires billing in order to succeed. I did, however, write a test program and verify that it worked.